+++ CHANGELOG.md
@@ -9,6 +9,14 @@ For help with updating to new Bytewax versions, please see the
 __Add any extra change notes here and we'll put them in the release
 notes on GitHub when we make a new release.__
 
+- Support for schema registries, through `bytewax.connectors.kafka.registry.RedpandaSchemaRegistry` and `bytewax.connectors.kafka.registry.ConfluentSchemaRegistry`.
+
+- Custom Kafka operators in `bytewax.connectors.kafka.operators`:
+  `input`, `output`, `deserialize_key`, `deserialize_value`, `deserialize`,
+  `serialize_key`, `serialize_value` and `serialize`.
+
+- *Breaking change* `KafkaSource` now emits a special `KafkaSourceMessage` to allow access to all data on consumed messages. `KafkaSink` now consumes `KafkaSinkMessage` to allow setting additional fields on produced messages.
+
 - Non-linear dataflows are now possible. Each operator method returns
   a handle to the `Stream`s it produces; add further steps via calling
   operator functions on those returned handles, not the root
@@ -35,6 +43,9 @@ notes on GitHub when we make a new release.__
 - *Breaking change* `fold` and `reduce` operators have been renamed to
   `fold_final` and `reduce_final`. They now only emit on EOF and are
   only for use in batch contexts.
+
+- *Breaking change* `batch` operator renamed to `collect`, so as to
+  not be confused with runtime batching. Behavior is unchanged.
 
 - *Breaking change* `output` operator does not forward downstream its
   items. Add operators on the upstream handle instead.